wta1/requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and serialization
pydantic>=2.5.0

# OpenAI integration
openai>=1.0.0
openai-agents>=0.3.0

# Vector database and embeddings
chromadb>=0.4.0
numpy>=1.24.0

# PDF and image processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Environment management
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
typing-extensions>=4.8.0
